Update to 1.18.0-rc.3

+++ Datastructures/IntDataMap2D.cs
@@ -83,7 +83,7 @@ namespace Vintagestory.API.Datastructures
             );
         }
 
-        public float GetUnpaddedIntLerpedForBlockPos(int x, int z, int regionSize)
+        public int GetUnpaddedIntLerpedForBlockPos(int x, int z, int regionSize)
         {
             int noiseSize = InnerSize;
